requirements.txt
pymupdf
pillow
opencv-python
numpy<2
tqdm
rich
pytesseract
paddleocr
torch
transformers
langgraph
openai
python-dotenv
